+++ Assets/Folder_LSR/Skills/MonsterSkillData.cs
@@ -14,27 +14,20 @@ public class MonsterSkillData : ScriptableObject, ISkillInfo
     [SerializeField] private ETier tier;
 
     [Header("전투용 스탯")]
-    [Tooltip("기본 데미지")]
-    [SerializeField] private float damage;
-
-    [Tooltip("공격력 계수")]
-    [SerializeField] private float coefficient;
-
-    [Tooltip("공격 범위")]
-    [SerializeField] private float range;
-
-    [Tooltip("쿨타임(초)")]
-    [SerializeField] private float coolTime;
+    [Tooltip("기본 데미지")][SerializeField] private int damage;
+    [Tooltip("공격력 계수")][SerializeField] private int coefficient;
+    [Tooltip("공격 범위")][SerializeField] private int range;
+    [Tooltip("쿨타임(초)")][SerializeField] private int coolTime;
 
     // ISkillInfo 
     public string Id => skillID;
     public string SkillName => skillName;
     public ESkillType SkillType => type;
     public ETier SkillTier => tier;
-    public float Damage => damage;
-    public float Coefficient => coefficient;
-    public float Range => range;
-    public float CoolTime => coolTime;
+    public int Damage => damage;
+    public int Coefficient => coefficient;
+    public int Range => range;
+    public int CoolTime => coolTime;
 
     // 몬스터 스킬은 해금X : 빈 리스트 반환
     public IReadOnlyList<ISkillInfo> UnlockNext => Array.Empty<ISkillInfo>();
